RIRB initialization, response from todo list: 0, output seperator lines added. check in docs if timer implementaation is essential.

+++ include/hda.h
@@ -85,7 +85,7 @@ void CORBInit();
 void RIRBInit();
 
 void HDASendCommand(u32 command);
-u32 HDAReadResponse();
+u64 HDAReadResponse();
 
 void HDACodecCommand(u32 verb, u32 payload);
 void HDAIdentifyCodecs();
